+++ Backend/app/dashboard/rider-dashboard/src/Domain/Action/RiderPlatform/Management/Merchant.hs
@@ -206,7 +206,7 @@ postMerchantConfigFailover merchantShortId opCity apiTokenInfo configName req = 
   transaction <- buildTransaction apiTokenInfo (Just req)
   T.withTransactionStoring transaction $ Client.callManagementAPI checkedMerchantId opCity (.merchantDSL.postMerchantConfigFailover) configName req
 
-postMerchantTicketConfigUpsert :: (Kernel.Types.Id.ShortId DM.Merchant -> Kernel.Types.Beckn.Context.City -> ApiTokenInfo -> Dashboard.Common.Merchant.UpsertTicketConfigReq -> Environment.Flow Dashboard.Common.Merchant.UpsertTicketConfigResp)
+postMerchantTicketConfigUpsert :: (Kernel.Types.Id.ShortId DM.Merchant -> Kernel.Types.Beckn.Context.City -> ApiTokenInfo -> Common.UpsertTicketConfigReq -> Environment.Flow Common.UpsertTicketConfigResp)
 postMerchantTicketConfigUpsert merchantShortId opCity apiTokenInfo req = do
   checkedMerchantId <- merchantCityAccessCheck merchantShortId apiTokenInfo.merchant.shortId opCity apiTokenInfo.city
   transaction <- buildTransaction apiTokenInfo (Just req)
